test(integration): consolidate 7 minimal-dep tests with mini seed
Merge cfg_gated_alloc, cpu_time, custom_bin_path, exit_in_non_main,
integration_frames, strict_lints, and workspace_member into a single
minimal_integration.rs binary with mini seed pre-population.
