--- tests/custom_bin_path.rs
@@ -1,102 +1,0 @@
-//! Test: piano build works when [[bin]] specifies a custom path.
-
-mod common;
-
-use std::fs;
-use std::path::Path;
-use std::process::Command;
-
-/// Create a project with `[[bin]] path = "src/myapp/main.rs"`.
-fn create_custom_bin_project(root: &Path) {
-    fs::create_dir_all(root.join("src").join("myapp")).unwrap();
-
-    fs::write(
-        root.join("Cargo.toml"),
-        r#"[package]
-name = "myapp"
-version = "0.1.0"
-edition = "2024"
-
-[[bin]]
-name = "myapp"
-path = "src/myapp/main.rs"
-"#,
-    )
-    .unwrap();
-
-    fs::write(
-        root.join("src").join("myapp").join("main.rs"),
-        r#"fn main() {
-    let result = compute();
-    println!("result: {result}");
-}
-
-fn compute() -> u64 {
-    (0..100).sum()
-}
-"#,
-    )
-    .unwrap();
-}
-
-#[test]
-fn custom_bin_path_produces_profiling_data() {
-    let tmp = tempfile::tempdir().unwrap();
-    let project = tmp.path().join("myapp");
-    create_custom_bin_project(&project);
-
-    let piano_bin = env!("CARGO_BIN_EXE_piano");
-    let manifest_dir = Path::new(env!("CARGO_MANIFEST_DIR"));
-    let runtime_path = manifest_dir.join("piano-runtime");
-
-    let output = Command::new(piano_bin)
-        .args(["build", "--fn", "compute", "--project"])
-        .arg(&project)
-        .arg("--runtime-path")
-        .arg(&runtime_path)
-        .output()
-        .expect("failed to run piano build");
-
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    let stdout = String::from_utf8_lossy(&output.stdout);
-
-    assert!(
-        output.status.success(),
-        "piano build failed:\nstderr: {stderr}\nstdout: {stdout}"
-    );
-
-    let binary_path = stdout.trim();
-    assert!(
-        Path::new(binary_path).exists(),
-        "built binary should exist at: {binary_path}"
-    );
-
-    // Run the instrumented binary.
-    let runs_dir = tmp.path().join("runs");
-    fs::create_dir_all(&runs_dir).unwrap();
-
-    let run_output = Command::new(binary_path)
-        .env("PIANO_RUNS_DIR", &runs_dir)
-        .output()
-        .expect("failed to run instrumented binary");
-
-    assert!(
-        run_output.status.success(),
-        "instrumented binary failed:\n{}",
-        String::from_utf8_lossy(&run_output.stderr)
-    );
-
-    let program_stdout = String::from_utf8_lossy(&run_output.stdout);
-    assert!(
-        program_stdout.contains("result: 4950"),
-        "program should produce correct output, got: {program_stdout}"
-    );
-
-    // Verify profiling data was actually written.
-    let run_file = common::largest_ndjson_file(&runs_dir);
-    let content = fs::read_to_string(&run_file).unwrap();
-    assert!(
-        content.contains("compute"),
-        "output should contain instrumented function name 'compute'"
-    );
-}
